Replaced old Guitar component
