--- src/components/Guitar/stores/focussedFret.ts
@@ -1,6 +1,0 @@
-import { writable } from 'svelte/store';
-
-export default writable({
-  string: 0,
-  fret: 0
-});
