--- src/components/Guitar/stores/index.ts
@@ -1,1 +1,0 @@
-export { default as focussedFret } from './focussedFret';
