--- src/components/Guitar/subcomponents/Fret/Fret.scss
@@ -1,71 +1,0 @@
-.fret {
-  --background-color: saddlebrown;
-  all: unset;
-  display: grid;
-  place-content: center;
-  flex: 1 1 0px;
-  height: 100%;
-  border: white solid 1px;
-  color: var(--text-dark-low-emphasis);
-  background-color: var(--background-color);
-  cursor: pointer;
-
-  &:hover {
-    background-color: hsl(from var(--background-color) h s 50);
-  }
-
-  &:first-child {
-    --background-color: black;
-    color: var(--text-light-low-emphasis-locked);
-  }
-
-  :global([data-theme="dark"]) & {
-    --background-color: black;
-
-    &:first-child {
-      --background-color: rgb(150, 150, 150);
-      color: var(--text-dark-low-emphasis-locked);
-
-      &:hover {
-        background-color: hsl(from var(--background-color) h s 70);
-      }
-    }
-  }
-
-  &__indicator {
-    position: relative;
-    z-index: 1;
-    display: grid;
-    place-content: center;
-    width: 1.5rem;
-    height: 1.5rem;
-    border-radius: 50%;
-  }
-
-  &[aria-selected="true"] > &__indicator {
-    background-color: rgb(255, 255, 255);
-    color: var(--text-dark-contrast-locked);
-    font-weight: 700;
-  }
-
-  &[aria-current="location"] {
-    position: relative;
-
-    &::after {
-      content: "";
-      position: absolute;
-      top: 0;
-      left: 0;
-      right: 0;
-      bottom: 0;
-      background-color: rgba(255, 255, 255, 0.3);
-    }
-
-    &:focus {
-      &::after {
-        outline: solid rgb(238, 212, 19) 0.25rem;
-        outline-offset: -0.25rem;
-      }
-    }
-  }
-}
