--- src/components/Guitar/subcomponents/Fret/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Fret.svelte';
