--- src/components/Guitar/subcomponents/String/String.scss
@@ -1,5 +1,0 @@
-.string {
-  display: flex;
-  height: 2rem;
-  width: 100%;
-}
