--- src/components/Guitar/subcomponents/String/index.ts
@@ -1,1 +1,0 @@
-export { default } from './String.svelte';
