+++ src/components/Guitar/subcomponents/index.ts
@@ -1,3 +1,1 @@
 export { default as Config } from './Config';
-export { default as Fret } from './Fret';
-export { default as String } from './String';
