--- src/components/Guitar/utils/focusFret.ts
@@ -1,15 +1,0 @@
-import { focussedFret } from '../stores';
-
-export default function focusFret({ string, fret }: { string: number, fret: number }) {
-  const target = document.querySelector<HTMLTableCellElement>(
-    `[data-row="${string}"][data-column="${fret}"]`
-  );
-  if (!target) {
-    return;
-  }
-  focussedFret.set({
-    string,
-    fret
-  });
-  target.focus();
-}
