--- src/components/Guitar/utils/index.ts
@@ -1,2 +1,0 @@
-export { default as focusFret } from './focusFret';
-export { default as handleKeyboardEvent } from './handleKeyboardEvent';
