--- src/components/NewGuitar/Guitar.scss
@@ -1,135 +1,0 @@
-.guitar {
-  all: unset;
-  display: flex;
-  flex-direction: column;
-  box-sizing: border-box;
-  gap: 0.125rem;
-  padding: 0.125rem;
-  width: 100%;
-  background-color: var(--text-light-high-emphasis-locked);
-}
-
-.string {
-  display: flex;
-  gap: 0.125rem;
-  width: 100%;
-  height: 2rem;
-}
-
-.fret {
-  position: relative;
-  display: block;
-  flex: 1 1 0px;
-  background-color: saddlebrown;
-  color: var(--text-dark-low-emphasis);
-  cursor: pointer;
-
-  &:first-child {
-    background-color: black;
-    color: var(--text-light-low-emphasis-locked);
-  }
-
-  &:hover {
-    &::after {
-      content: "";
-      position: absolute;
-      top: 0;
-      left: 0;
-      right: 0;
-      bottom: 0;
-      background-color: rgba(238, 212, 19, 0.2);
-    }
-  }
-
-  &:focus-within {
-    outline: solid rgb(238, 212, 19) 0.25rem;
-    outline-offset: -0.25rem;
-
-    &::after {
-      content: "";
-      position: absolute;
-      top: 0;
-      left: 0;
-      right: 0;
-      bottom: 0;
-      background-color: rgba(238, 212, 19, 0.3);
-    }
-  }
-
-  :global([data-theme="dark"]) & {
-    background-color: black;
-
-    &:first-child {
-      background-color: rgb(90, 90, 90);
-      color: var(--text-light-low-emphasis-locked);
-    }
-  }
-  
-  &__label {
-    position: absolute;
-    display: grid;
-    place-content: center;
-    z-index: 1;
-    left: 50%;
-    top: 50%;
-    width: 1.5rem;
-    height: 1.5rem;
-    font-size: 1rem;
-    border-radius: 50%;
-    transform: translate(-50%, -50%);
-  }
-
-  &__input {
-    position: absolute;
-    opacity: 0;
-    height: 0;
-    width: 0;
-    cursor: pointer;
-  }
-
-  &__input:checked + &__label {
-    background-color: white;
-    color: var(--text-dark-contrast-locked);
-    font-weight: bold;
-  }
-
-  &.isRoot > &__input:checked + &__label {
-    background-color: red;
-    color: var(--text-light-contrast-locked);
-    font-weight: bold;
-  }
-}
-
-.fret-markers {
-  box-sizing: border-box;
-  display: flex;
-  gap: 0.125rem;
-  padding: 0 0.125rem;
-  width: 100%;
-}
-
-.fret-marker {
-  flex: 1 1 0px;
-  
-  &.isHighlighted {
-    border-color: var(--text-dark-high-emphasis);
-  }
-  
-  &--top {
-    display: grid;
-    place-content: center;
-    padding: 0.25rem;
-    border-bottom: 0.5rem solid transparent;
-    color: var(--text-dark-low-emphasis);
-
-    &.isHighlighted {
-      color: var(--text-dark-high-emphasis);
-      font-weight: bold;
-    }
-  }
-
-  &--bottom {
-    flex: 1 1 0px;
-    border-top: 0.5rem solid transparent;
-  }
-}
