--- src/components/NewGuitar/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Guitar.svelte';
