--- src/components/NewGuitar/subcomponents/Config/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Config.svelte';
