--- src/components/NewGuitar/subcomponents/index.ts
@@ -1,1 +1,0 @@
-export { default as Config } from './Config';
